Add files via upload
This is the data file which calculates the pluralistic ignorance for the 5 AMI questions. It is a different file becuase the estimates for the English public answer and the answers participants gave themselves are stacked ontop of one another, so that I could regress against a dummy variable, 1 = your own answer, 0 = estimate of the English public's answer.