Add files via upload
Running this do file generates the values for the pluralistic ignorance between the answers participants gave for their own beliefs in response to Q1-5 on the AMI items, and also what they estimated the public would answer in response to the same questions. It pulls from the dataset: "dataset PLUR IGNO with variables", which is a STATA dataset.